+++ CHANGELOG.md
@@ -10,13 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2023-02-20
 
 ### Added
+
 - Implement no_std environment support for all features except streaming
 
 ### Removed
+
 - Set default-features to false
 - Removed unnecessary thiserror package
 
 ### Changed
+
 - Refactored custom Error enum to not depend on thiserror and to only use std for the stream feature
 - Use alloc and core in place of std where possible