+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-09-15
+
 ### Added
 
 - New default `std` feature enabling `thiserror` crate to handle `std::error::Error` implementation on `Error` enum
@@ -106,7 +108,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the library
 
-[Unreleased]: https://github.com/monero-rs/base58-monero/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/monero-rs/base58-monero/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/monero-rs/base58-monero/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/monero-rs/base58-monero/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/monero-rs/base58-monero/compare/v0.3.2...v1.0.0
 [0.3.2]: https://github.com/monero-rs/base58-monero/compare/v0.3.1...v0.3.2